src/consumer/requirements.txt
kafka-python==2.0.2
pymongo==4.6.1
psycopg2-binary==2.9.9
python-dotenv==1.0.0
pytz==2024.1